+++ api/v1/routes/event.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const axios = require("axios");
 const cache = require("../scripts/cache");
-const config = require("../../../config");
+
 
 router.get("/", (req, res) => {
   res.status(501).json({ error: "not implemented" });
@@ -17,52 +17,22 @@ async function official_events(req, res) {
   let from = parseInt(req.query.from);
   let to = parseInt(req.query.to);
 
-  if (/*cache.get(key)*/ false) {
-    var data = cache.get(key);
-  } else {
-    const urls = [
-      "https://api.brussels:443/api/agenda/0.0.1/events/search?name=fiets",
-      "https://api.brussels:443/api/agenda/0.0.1/events/search?name=v%C3%A9lo",
-      "https://api.brussels:443/api/agenda/0.0.1/events/search?name=cycli",
-    ];
-
-    let jsons = urls.map(async function (item) {
-      try {
-        let json = await fetch(item);
-        return json.data;
-      } catch (err) {
-        console.log(err);
-        return null;
-      }
-    });
-
-    let jsons_nopromise = await Promise.all(jsons);
-
-    let all_events = {
-      events: [],
-    };
-
-    jsons_nopromise.forEach((item) => {
-      if (item.response.resultCount > 0) {
-        item.response.results.event.forEach((elem) => {
-          all_events.events.push(elem);
-        });
-      }
-    });
-
+  if (cache.get(key)) {
+    let all_events = cache.get(key);
     all_events.events = filterByDate(all_events.events, from, to);
-
     return res.status(200).json(all_events);
+  } else {
+    return res.status(500).json({ error: "internal server error" });
   }
 }
 
 async function official_event_by_id(req,res){
     try {
-        var json = (await fetch(`https://api.brussels:443/api/agenda/0.0.1/events/${req.params.id}`)).data;
-      } catch (err) {
-          console.log(err);
-        return res.status(500).json({ error: "internal server error" });
-      }
+      var json = (await fetch(`https://api.brussels:443/api/agenda/0.0.1/events/${req.params.id}`)).data;
+    } catch (err) {
+      console.log(err);
+      return res.status(500).json({ error: "internal server error" });
+    }
 
       return res.status(200).json(json);
 }
@@ -91,18 +61,5 @@ function filterByDate(events, from, to) {
   return filtered_events;
 }
 
-function fetch(url) {
-  const options = {
-    headers: {
-      Authorization: getAPIKey(),
-      Accept: "application/json",
-    },
-  };
-  return axios.get(url, options);
-}
-
-function getAPIKey() {
-  return config.env.APIBRUSSELS_API_KEY;
-}
 
 module.exports = router;
