+++ api/v1/scripts/cache.js
@@ -2,18 +2,26 @@ const cache = require("memory-cache");
 
 const dataFetcher = require("../scripts/datafetcher");
 
-const OPENSTREETMAP_REFRESH = 86400000; // one day
-const VILLO_STATION_REFRESH = 3600000; // one hour
+const OPENSTREETMAP_REFRESH = 86400000;     // one day
+const VILLO_STATION_REFRESH = 3600000;      // one hour
+const OFFICIAL_EVENTS_REFRESH = 3600000     // one hour
 
 module.exports = {
 
     init: async function() {
+        // Points of interests
+        /*
         await refreshBicycleParking();
+
         await refreshVilloStation();
         await refreshCompressedAir();
         await refreshBicycleRepairStations();
         await refreshBicycleShop();
+
         await refreshDrinkingWater();
+        */
+        // Events
+        await refreshOfficialEvents();
     },
     add: function (key, data) {
         cache.put(key, data);
@@ -46,4 +54,8 @@ async function refreshBicycleShop() {
 
 async function refreshDrinkingWater() {
     await cache.put("drinking_water", await dataFetcher.fetchDrinkingWater(), OPENSTREETMAP_REFRESH, refreshDrinkingWater);
+}
+
+async function refreshOfficialEvents() {
+    await cache.put("official_events", await dataFetcher.fetchOfficialEvents(), OFFICIAL_EVENTS_REFRESH, refreshOfficialEvents);
 }
